Use "." instead of tabs when flattening nested keys
I'm making this "breaking" change because tabs seem to be oddly escaped
in different situations. It's safer to use something like a "." instead.
Also, it's easier to read a key separated by dots instead of tabs.

There is a risk of key collision if people used keys with dots in them,
but there would also be a key collision before as well with tabs.

Also added `test-py` target to `test-all` in Makefile since it's so much
faster than running `make test`